Refactoring of common functions to AuthenticationViewModel.kt and ProfileViewModel.kt
